+++ src/viur/datastore/config.py
@@ -2,5 +2,10 @@
 	This file just holds some configuration variables that will influence
 	the behaviour of this library.
 """
-traceQueries = False  # If set, we'll log each query we run
-SkeletonInstanceRef = None  # A reference to the skeleton container of ViUR. Unless set, fetch() and getSkel() will fail
+conf = {
+	# If set, we'll log each query we run
+	"traceQueries": False,
+
+	# A reference to the skeleton container of ViUR. Unless set, fetch() and getSkel() will fail
+	"SkeletonInstanceRef": None,
+}